additions: Player hp count on UI changes: new damage system (not based on projectile data)

+++ src/main/kotlin/dev/zorsh/zorshDeltarune/battle/INeverlandBattle.kt
@@ -21,7 +21,7 @@ interface INeverlandBattle {
 
     fun getBBLocation(): Pair<Float, Float>
 
-    fun createProjectile(px: Float, py: Float, projectileData: ProjectileData, afterCreated: (PlayerUICanvas, String, () -> Unit) -> Unit)
+    fun createProjectile(px: Float, py: Float, projectileData: ProjectileData, afterCreated: (PlayerUICanvas, String, (Int) -> Unit) -> Unit)
 
     fun removePlayer(dPlayer: DeltarunePlayer)
 }
